+++ src/ft_vector_draw.c
@@ -6,22 +6,22 @@
 /*   By: msamhaou <msamhaou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/21 17:24:33 by msamhaou          #+#    #+#             */
-/*   Updated: 2023/07/11 05:23:38 by msamhaou         ###   ########.fr       */
+/*   Updated: 2023/07/12 05:59:38 by msamhaou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-float	*ft_hypo_calc(t_vertex *a, t_vertex *b)
+double	*ft_hypo_calc(t_vertex *a, t_vertex *b)
 {
-	float	*hypo;
-	float	x;
-	float	y;
-	// float	deg;
+	double	*hypo;
+	double	x;
+	double	y;
+	// double	deg;
 
 	x = (b->x - a->x);
 	y = (a->y - b->y);
-	hypo = malloc(sizeof(float) * 2);
+	hypo = malloc(sizeof(double) * 2);
 	hypo[0] = sqrt((x * x) + (y * y));
 	hypo[1] = asin(y / hypo[0]);
 	return (hypo);
@@ -38,13 +38,13 @@ void	ft_swap_vect(t_vertex **a, t_vertex **b)
 
 void	ft_vect_draw(t_vertex *a, t_vertex *b, int color, t_data *data)
 {
-	float	i;
-	float	*hypo;
-	float	*hypo_max;
-	// float	deg;
-	// float	adj;
-	float	x;
-	float	y;
+	double	i;
+	double	*hypo;
+	double	*hypo_max;
+	// double	deg;
+	// double	adj;
+	double	x;
+	double	y;
 
 	i = 0;
 	hypo = ft_hypo_calc(a, b);
